--- MobileApp/app/src/main/java/com/example/fe_clientapp/Login.java
@@ -1,31 +1,0 @@
-package com.example.fe_clientapp;
-
-import android.content.Intent;
-import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class Login extends AppCompatActivity {
-  private Button button;
-
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setContentView(R.layout.login);
-
-    button = (Button) findViewById(R.id.loginBtn);
-    button.setOnClickListener(new View.OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        openMain();
-      }
-    });
-  }
-
-  public void openMain() {
-    Intent intent = new Intent(this, MainActivity.class);
-    startActivity(intent);
-    overridePendingTransition(R.anim.slide_in_right, R.anim.slide_out_left);
-  }
-}
