--- MobileApp/app/src/main/java/com/example/fe_clientapp/ui/login/LoginFragment.java
@@ -1,28 +1,0 @@
-package com.example.fe_clientapp.ui.login;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
-
-import com.example.fe_clientapp.R;
-
-public class LoginFragment extends Fragment {
-
-  private LoginViewModel loginViewModel;
-
-  public View onCreateView(@NonNull LayoutInflater inflater,
-                           ViewGroup container, Bundle savedInstanceState) {
-    loginViewModel =
-      ViewModelProviders.of(this).get(LoginViewModel.class);
-    View root = inflater.inflate(R.layout.login, container, false);
-    return root;
-  }
-}
