--- MobileApp/app/src/main/java/com/example/fe_clientapp/ui/login/LoginViewModel.java
@@ -1,19 +1,0 @@
-package com.example.fe_clientapp.ui.login;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class LoginViewModel extends ViewModel {
-
-  private MutableLiveData<String> mText;
-
-  public LoginViewModel() {
-    mText = new MutableLiveData<>();
-    mText.setValue("This is login fragment");
-  }
-
-  public LiveData<String> getText() {
-    return mText;
-  }
-}
